+++ R/fuj-package.R
@@ -16,6 +16,8 @@ NULL
 #' @export
 # nolint next: object_name_linter.
 op.fuj <- list(
+  fuj.vap.progress = FALSE,
+  fuj.vap.indexed_errors = FALSE,
   fuj.verbose = NULL,
   fuj.verbose.fill = FALSE,
   fuj.verbose.label = "verbose: "